비트연산 프로토콜 작업중(send나 recv했을때 이상한값이 들어오는중) - 수정진행중

+++ Server/MonsterTime.cpp
@@ -5,11 +5,11 @@ void MonsterTime::Start_Time()
 	start = clock();
 }
 
-int MonsterTime::End_Time()
+double MonsterTime::End_Time()
 {
 	end = clock();
 
-	result = (end - start) / CLOCKS_PER_SEC;;
+	result = (double)(end - start) / CLOCKS_PER_SEC;;
 
 	return 	result; ;
 }
